+++ lib/typescript/commonjs/core/createCollectionModel.d.ts
@@ -1,7 +1,8 @@
-import type { CollectionModel, CreateCollectionModelConfig } from '../types';
+import type { CollectionModel, CreateCollectionModelFieldsConfig, CreateCollectionModelNormalizeConfig, ModelFieldSpecs, ModelStoredFromFields } from '../types';
 /** Create a collection model from a persistent collection and normalizer. */
 export declare function createCollectionModel<TInput, TStored extends {
     id: string;
     updatedAt?: string | null;
-}, TExt extends Record<string, unknown> = {}>(config: CreateCollectionModelConfig<TInput, TStored, TExt>): CollectionModel<TInput, TStored> & TExt;
+}, TExt extends Record<string, unknown> = {}>(config: CreateCollectionModelNormalizeConfig<TInput, TStored, TExt>): CollectionModel<TInput, TStored> & TExt;
+export declare function createCollectionModel<TFields extends ModelFieldSpecs, TExt extends Record<string, unknown> = {}>(config: CreateCollectionModelFieldsConfig<TFields, TExt>): CollectionModel<unknown, ModelStoredFromFields<TFields>> & TExt;
 //# sourceMappingURL=createCollectionModel.d.ts.map